Backwards compatibility and module retention after installation
